Unificar smoke tests de SolicitarProgramacionTurno en una sola clase
- Fusionar SolicitarProgramacionTurnoSmokeTests y SolicitarProgramacionTurnoSbSmokeTests
  en una sola clase que recibe ApiFixture y ServiceBusFixture
- Eliminar test redundante DebeRetornar202_CuandoPayloadEsValido (subconjunto
  de DebePublicar...CuandoSolicitudEsAceptada)
- Enriquecer test 409 para consumir los 2 eventos publicados por el primer POST,
  verificar fechas y evitar mensajes huerfanos en la suscripcion
- Agregar PurgeAsync y Assert.SkipWhen al test 409

Closes #99
